01-simple-react-agent/requirements.txt
fastapi
uvicorn[standard]
chainlit
httpx
langchain
langchain-groq
python-dotenv
psycopg2[binary]
langsmith